+++ Assets/Scripts/Item/ItemDataWeapon.cs
@@ -22,15 +22,15 @@ public class ItemDataWeapon : ItemDataEquipment
     }
     
     [SerializeField] private AttackStyle _atkStyle;
-    [SerializeField] private float _minAtk;
-    [SerializeField] private float _maxAtk;
+    [SerializeField] private int _minAtk;
+    [SerializeField] private int _maxAtk;
     [SerializeField] private float _atkRange;
     [SerializeField] private float _atkSpeed;
     [SerializeField] private int _maxAtkCount;
 
     public AttackStyle AtkStyle => _atkStyle;
-    public float MinAtk => _minAtk;
-    public float MaxAtk => _maxAtk;
+    public int MinAtk => _minAtk;
+    public int MaxAtk => _maxAtk;
     public float AtkRange => _atkRange;
     public float AtkSpeed => _atkSpeed;
     public int MaxAtkCount => _maxAtkCount;
